+++ src/Game/Text/el/HelpText.el.sjson.lua
@@ -171,7 +171,7 @@ local newData = {
 	-- #region Credits
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Craft_NikkelM",
-	-- 	DisplayName = "a mod by",
+	-- 	DisplayName = "A Mod By",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Name_NikkelM",
@@ -179,11 +179,11 @@ local newData = {
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Info_Special_Shoutouts",
-	-- 	DisplayName = "modding community contributors",
+	-- 	DisplayName = "Modding Community Contributors",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Craft_iDeath",
-	-- 	DisplayName = "hell2modding",
+	-- 	DisplayName = "Hell2Modding",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Name_iDeath",
@@ -191,7 +191,7 @@ local newData = {
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Craft_burn",
-	-- 	DisplayName = "artwork",
+	-- 	DisplayName = "Art",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Name_burn",
@@ -199,11 +199,19 @@ local newData = {
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Craft_Magic",
-	-- 	DisplayName = "modutil",
+	-- 	DisplayName = "Modutil",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Name_Magic",
 	-- 	DisplayName = "Magic_Gonads",
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Craft_zannc",
+	-- 	DisplayName = "Keepsake Port",
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Name_zannc",
+	-- 	DisplayName = "zannc",
 	-- },
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_InGameCredits_Info_Community",